+++ src/main/java/net/jrodolfo/java_evolution/java06/compiler_api/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Explanatory module for the Java 6 Compiler API.
+ * Executable examples for the Java 6 Compiler API.
  */
 package net.jrodolfo.java_evolution.java06.compiler_api;